Created BaseApiController, and filtered droplist according to the building where they belong

+++ Droplist.api/Controllers/DroplistsController.cs
@@ -7,30 +7,37 @@ using System.Net;
 using System.Web.Http;
 using System.Web.Http.Description;
 using Droplist.api.Models;
+using Droplist.api.Utility;
 
 namespace Droplist.api.Controllers
 {
-    public class DroplistsController : ApiController
+    public class DroplistsController : BaseApiController
     {
-        private DroplistDataContext db = new DroplistDataContext();
+        //private DroplistDataContext db = new DroplistDataContext();
 
         // GET: api/Droplists
         public IHttpActionResult GetDroplists()
         {
-            var resultSet = db.Droplists.Select(droplist => new
-            {
-                droplist.DroplistId,
-                droplist.BuildingId,
-                droplist.DroplistName,
-                droplist.CreatedOnDate,
-                droplist.StockerId,
-                droplist.DriverId,
-                droplist.SectionId,
-                DriverName = droplist.Driver.FirstName + " " + droplist.Driver.LastName,
-                StockerName = droplist.Stocker.FirstName + " " + droplist.Stocker.LastName,
-                DepartmentName = droplist.Section.Department.DepartmentName,
-                SectionName = droplist.Section.SectionName
-            });
+            if (CurrentUser == null) {
+                return Unauthorized();
+            }
+
+            var resultSet = db.Droplists
+                .Where(d=> d.BuildingId == CurrentUser.Employee.BuildingId) //Match userId of logged user
+                .Select(droplist => new
+                {
+                    droplist.DroplistId,
+                    droplist.BuildingId,
+                    droplist.DroplistName,
+                    droplist.CreatedOnDate,
+                    droplist.StockerId,
+                    droplist.DriverId,
+                    droplist.SectionId,
+                    DriverName = droplist.Driver.FirstName + " " + droplist.Driver.LastName,
+                    StockerName = droplist.Stocker.FirstName,
+                    DepartmentName = droplist.Section.Department.DepartmentName,
+                    SectionName = droplist.Section.SectionName
+                });
 
             return Ok(resultSet);
         }
@@ -39,36 +46,41 @@ namespace Droplist.api.Controllers
         [ResponseType(typeof(Models.Droplist))]
         public IHttpActionResult GetDroplist(int id)
         {
+            if (CurrentUser == null)
+            {
+                return Unauthorized();
+            }
+
             Models.Droplist droplist = db.Droplists.Find(id);
             if (droplist == null)
             {
                 return NotFound();
             }
 
-			var droplistItems = db.DroplistItems
-				.Where(d => d.DroplistId == id)
-				.Select(d => new
-				{
-					product = new
-					{
-						productId = d.Product.ProductId,
-						itemNumber = d.Product.ItemNumber,
-						description = d.Product.Description
-					},
-					droplistId = d.DroplistId,
-					aisleNumber = d.AisleNumber,
-					aisleRow = d.AisleRow,
-					aisleColumn = d.AisleColumn,
-					completed = d.Completed,
-					rejected = d.Rejected,
-					quantity = d.Quantity,
-					droplistItemId = d.DroplistItemId
-				})
+            var droplistItems = db.DroplistItems
+                .Where(d => d.DroplistId == id)
+                .Select(d => new
+                {
+                    product = new
+                    {
+                        productId = d.Product.ProductId,
+                        itemNumber = d.Product.ItemNumber,
+                        description = d.Product.Description
+                    },
+                    droplistId = d.DroplistId,
+                    aisleNumber = d.AisleNumber,
+                    aisleRow = d.AisleRow,
+                    aisleColumn = d.AisleColumn,
+                    completed = d.Completed,
+                    rejected = d.Rejected,
+                    quantity = d.Quantity,
+                    droplistItemId = d.DroplistItemId
+                })
 
-				.OrderBy(d => d.aisleNumber)
-				.ThenBy(d => d.aisleColumn)
-				.ThenBy(d => d.aisleRow);
-                
+                .OrderBy(d => d.aisleNumber)
+                .ThenBy(d => d.aisleColumn)
+                .ThenBy(d => d.aisleRow);
+
 
             var resultSet = new
             {
@@ -79,7 +91,7 @@ namespace Droplist.api.Controllers
                 droplist.StockerId,
                 droplist.DriverId,
                 droplist.SectionId,
-                DriverName = (droplist.Driver!= null) ? droplist.Driver.FirstName + " " + droplist.Driver.LastName : null,
+                DriverName = (droplist.Driver != null) ? droplist.Driver.FirstName + " " + droplist.Driver.LastName : null,
                 //StockerName = droplist.Stocker.FirstName + " " + droplist.Stocker.LastName,
                 DepartmentName = droplist.Section.Department.DepartmentName,
                 SectionName = droplist.Section.SectionName,
@@ -92,6 +104,11 @@ namespace Droplist.api.Controllers
         [ResponseType(typeof(void))]
         public IHttpActionResult PutDroplist(int id, Models.Droplist droplist)
         {
+            if (CurrentUser == null)
+            {
+                return Unauthorized();
+            }
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -137,21 +154,21 @@ namespace Droplist.api.Controllers
                 }
                 else
                 {
-					var dbDroplistItem = db.DroplistItems.Find(droplistItem.DroplistItemId);
+                    var dbDroplistItem = db.DroplistItems.Find(droplistItem.DroplistItemId);
 
-					// save the properties you want to set
-					// aisle
-					dbDroplistItem.AisleNumber = droplistItem.AisleNumber;
-					// row
-					dbDroplistItem.AisleRow = droplistItem.AisleRow;
-					// column
-					dbDroplistItem.AisleColumn = droplistItem.AisleColumn;
-					// productid 
-					dbDroplistItem.ProductId = droplistItem.Product.ProductId;
-					// quantity
-					dbDroplistItem.Quantity = droplistItem.Quantity;
+                    // save the properties you want to set
+                    // aisle
+                    dbDroplistItem.AisleNumber = droplistItem.AisleNumber;
+                    // row
+                    dbDroplistItem.AisleRow = droplistItem.AisleRow;
+                    // column
+                    dbDroplistItem.AisleColumn = droplistItem.AisleColumn;
+                    // productid 
+                    dbDroplistItem.ProductId = droplistItem.Product.ProductId;
+                    // quantity
+                    dbDroplistItem.Quantity = droplistItem.Quantity;
 
-					db.Entry(dbDroplistItem).State = EntityState.Modified;
+                    db.Entry(dbDroplistItem).State = EntityState.Modified;
                 }
 
                 db.SaveChanges();
@@ -164,6 +181,11 @@ namespace Droplist.api.Controllers
         [ResponseType(typeof(Models.Droplist))]
         public IHttpActionResult PostDroplist(Models.Droplist droplist)
         {
+            if (CurrentUser == null)
+            {
+                return Unauthorized();
+            }
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -182,50 +204,51 @@ namespace Droplist.api.Controllers
 
             foreach (var di in droplist.DroplistItems)
             {
-				if (di.ProductId != 0)
-				{
-					var dbdi = new Models.DroplistItem
-					{
-						AisleColumn = di.AisleColumn,
-						AisleNumber = di.AisleNumber,
-						AisleRow = di.AisleRow,
-						Completed = di.Completed,
-						DroplistId = di.DroplistId,
-						DroplistItemId = di.DroplistItemId,
-						ProductId = di.ProductId,
-						Quantity = di.Quantity,
-						Rejected = di.Rejected
-					};
+                if (di.ProductId != 0)
+                {
+                    var dbdi = new Models.DroplistItem
+                    {
+                        AisleColumn = di.AisleColumn,
+                        AisleNumber = di.AisleNumber,
+                        AisleRow = di.AisleRow,
+                        Completed = di.Completed,
+                        DroplistId = di.DroplistId,
+                        DroplistItemId = di.DroplistItemId,
+                        ProductId = di.ProductId,
+                        Quantity = di.Quantity,
+                        Rejected = di.Rejected
+                    };
 
-					dbDroplist.DroplistItems.Add(dbdi);
-				}
-				else
-				{
-					var product = new Product {
-						ItemNumber = di.Product.ItemNumber,
-						Description = di.Product.Description,
-						Price = 0
-					
-					};
+                    dbDroplist.DroplistItems.Add(dbdi);
+                }
+                else
+                {
+                    var product = new Product
+                    {
+                        ItemNumber = di.Product.ItemNumber,
+                        Description = di.Product.Description,
+                        Price = 0
 
-					db.Products.Add(product);
-					db.SaveChanges();
+                    };
 
-					var dbdi = new Models.DroplistItem
-					{
-						AisleColumn = di.AisleColumn,
-						AisleNumber = di.AisleNumber,
-						AisleRow = di.AisleRow,
-						Completed = di.Completed,
-						DroplistId = di.DroplistId,
-						DroplistItemId = di.DroplistItemId,
-						ProductId = product.ProductId,
-						Quantity = di.Quantity,
-						Rejected = di.Rejected
-					};
+                    db.Products.Add(product);
+                    db.SaveChanges();
 
-					dbDroplist.DroplistItems.Add(dbdi);
-				}      
+                    var dbdi = new Models.DroplistItem
+                    {
+                        AisleColumn = di.AisleColumn,
+                        AisleNumber = di.AisleNumber,
+                        AisleRow = di.AisleRow,
+                        Completed = di.Completed,
+                        DroplistId = di.DroplistId,
+                        DroplistItemId = di.DroplistItemId,
+                        ProductId = product.ProductId,
+                        Quantity = di.Quantity,
+                        Rejected = di.Rejected
+                    };
+
+                    dbDroplist.DroplistItems.Add(dbdi);
+                }
             }
 
             db.Droplists.Add(dbDroplist);
@@ -233,7 +256,7 @@ namespace Droplist.api.Controllers
 
             return CreatedAtRoute("DefaultApi", new { id = dbDroplist.DroplistId }, new
             {
-				dbDroplist.DroplistId,
+                dbDroplist.DroplistId,
                 droplist.BuildingId,
                 droplist.DroplistName,
                 droplist.CreatedOnDate,
@@ -247,16 +270,21 @@ namespace Droplist.api.Controllers
         [ResponseType(typeof(Models.Droplist))]
         public IHttpActionResult DeleteDroplist(int id)
         {
+            if (CurrentUser == null)
+            {
+                return Unauthorized();
+            }
+
             Models.Droplist droplist = db.Droplists.Find(id);
             if (droplist == null)
             {
                 return NotFound();
             }
-			var droplistItems = db.DroplistItems.Where(x => x.DroplistId == id).ToList();
-			for (int i = 0; i < droplistItems.Count; i++)
-			{
-				db.DroplistItems.Remove(droplistItems[i]);
-			}
+            var droplistItems = db.DroplistItems.Where(x => x.DroplistId == id).ToList();
+            for (int i = 0; i < droplistItems.Count; i++)
+            {
+                db.DroplistItems.Remove(droplistItems[i]);
+            }
 
             db.Droplists.Remove(droplist);
             db.SaveChanges();
